requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.3
yt-dlp>=2024.12.1
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
psutil==5.9.8
nest-asyncio==1.6.0
ytmusicapi==1.5.3
google-generativeai
